Move process memory related code into :core
