+++ subprojects/core/src/main/java/org/gradle/process/internal/health/memory/MemoryInfo.java
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-package org.gradle.launcher.daemon.server.health.memory;
+package org.gradle.process.internal.health.memory;
 
 import org.gradle.internal.os.OperatingSystem;
-
-import static org.gradle.launcher.daemon.server.health.memory.MBeanAttributeProvider.getMbeanAttribute;
 
 public class MemoryInfo implements MemoryStatus {
 
@@ -52,7 +50,7 @@ public class MemoryInfo implements MemoryStatus {
      */
     @Override
     public long getTotalPhysicalMemory() {
-        return getMbeanAttribute("java.lang:type=OperatingSystem", "TotalPhysicalMemorySize", Long.class);
+        return MBeanAttributeProvider.getMbeanAttribute("java.lang:type=OperatingSystem", "TotalPhysicalMemorySize", Long.class);
     }
 
     /**
