+++ subprojects/core/src/test/groovy/org/gradle/process/internal/health/memory/MemoryInfoTest.groovy
@@ -13,8 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.launcher.daemon.server.health.memory
 
+
+package org.gradle.process.internal.health.memory
+
+import org.gradle.process.internal.health.memory.MemoryInfo
 import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
 import org.gradle.util.Requires
 import org.gradle.util.TestPrecondition
