+++ subprojects/core/src/test/groovy/org/gradle/process/internal/health/memory/VmstatAvailableMemoryTest.groovy
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package org.gradle.launcher.daemon.server.health.memory
+package org.gradle.process.internal.health.memory
 
+import org.gradle.process.internal.health.memory.VmstatAvailableMemory
 import spock.lang.Specification
 
 class VmstatAvailableMemoryTest extends Specification {
