+++ subprojects/launcher/src/main/java/org/gradle/launcher/daemon/server/health/DaemonHealthStats.java
@@ -22,9 +22,9 @@ import org.gradle.launcher.daemon.server.health.gc.GarbageCollectionInfo;
 import org.gradle.launcher.daemon.server.health.gc.GarbageCollectionMonitor;
 import org.gradle.launcher.daemon.server.health.gc.GarbageCollectionStats;
 import org.gradle.launcher.daemon.server.health.gc.GarbageCollectorMonitoringStrategy;
-import org.gradle.launcher.daemon.server.health.memory.MemoryInfo;
-import org.gradle.launcher.daemon.server.health.memory.MemoryStatus;
 import org.gradle.launcher.daemon.server.stats.DaemonRunningStats;
+import org.gradle.process.internal.health.memory.MemoryInfo;
+import org.gradle.process.internal.health.memory.MemoryStatus;
 
 import java.util.concurrent.ScheduledExecutorService;
