+++ subprojects/launcher/src/main/java/org/gradle/tooling/internal/provider/LauncherServices.java
@@ -28,9 +28,9 @@ import org.gradle.internal.service.ServiceRegistry;
 import org.gradle.internal.service.scopes.GradleUserHomeScopePluginServices;
 import org.gradle.internal.service.scopes.GradleUserHomeScopeServiceRegistry;
 import org.gradle.internal.service.scopes.PluginServiceRegistry;
-import org.gradle.launcher.daemon.server.health.memory.DefaultMemoryResourceManager;
-import org.gradle.launcher.daemon.server.health.memory.MemoryResourceManager;
-import org.gradle.launcher.daemon.server.health.memory.MemoryStatusBroadcaster;
+import org.gradle.process.internal.health.memory.DefaultMemoryResourceManager;
+import org.gradle.process.internal.health.memory.MemoryResourceManager;
+import org.gradle.process.internal.health.memory.MemoryStatusBroadcaster;
 import org.gradle.launcher.exec.BuildExecuter;
 import org.gradle.launcher.exec.ChainingBuildActionRunner;
 import org.gradle.launcher.exec.InProcessBuildActionExecuter;
