+++ subprojects/launcher/src/test/groovy/org/gradle/launcher/daemon/server/health/DaemonHealthStatsTest.groovy
@@ -19,8 +19,8 @@ package org.gradle.launcher.daemon.server.health
 import org.gradle.launcher.daemon.server.health.gc.GarbageCollectionInfo
 import org.gradle.launcher.daemon.server.health.gc.GarbageCollectionMonitor
 import org.gradle.launcher.daemon.server.health.gc.GarbageCollectionStats
-import org.gradle.launcher.daemon.server.health.memory.MemoryInfo
 import org.gradle.launcher.daemon.server.stats.DaemonRunningStats
+import org.gradle.process.internal.health.memory.MemoryInfo
 import spock.lang.Specification
 
 class DaemonHealthStatsTest extends Specification {
