+++ subprojects/launcher/src/test/groovy/org/gradle/launcher/daemon/server/health/LowMemoryDaemonExpirationStrategyTest.groovy
@@ -19,7 +19,7 @@ package org.gradle.launcher.daemon.server.health
 
 import com.google.common.base.Strings
 import org.gradle.launcher.daemon.server.expiry.DaemonExpirationResult
-import org.gradle.launcher.daemon.server.health.memory.MemoryStatus
+import org.gradle.process.internal.health.memory.MemoryStatus
 import spock.lang.Specification
 
 import static org.gradle.launcher.daemon.server.expiry.DaemonExpirationStatus.DO_NOT_EXPIRE
